# gotham_like/docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Tessera Investigative Workbench",
  "description": "Local-first investigative intelligence & decision-support platform. Outputs are decision support: system inferences and analytical signals are never presented as established facts.",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/system/errors": {
   "get": {
    "tags": [
     "system"
    ],
    "summary": "Recent Errors",
    "operationId": "recent_errors_api_system_errors_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Recent Errors Api System Errors Get"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  }
 },
 "components": {
...
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "scheme": "bearer"
   }
  }
 }
}